# openapi.yaml
openapi: 3.0.0
info:
  version: '1.1'
  title: Ruokavälitys API
  description: API for both, managing the RV system and it's inventory, and browsing and purchasing products.
...
servers:
  - url: https://rv-backend-dev.herokuapp.com/
    description: Staging environment
  - url: https://rv-backend.herokuapp.com/
    description: Production environment
...
components:
  securitySchemes:
    normalAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description:
        A JWT access token used for authenticating with the \"end user facing\" portition of the API.
        Can be obtained with any valid credentials from route `/api/v1/authenticate`.
    adminAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description:
        A JWT access token used for authenticating with the administrations portition of the API.
        Can be obtained with administrator credentials from route `/api/v1/admin/authenticate`.
  schemas:
...
    Product:
      type: object
      required: 
        - barcode
        - name
        - category
        - sellPrice
        - stock
      properties: 
        barcode: 
          type: string
          example: "6430039222212"
          description: Barcode on the product packaging. Used to uniquely identify the product.
        name: 
          type: string
          example: Kombucha bakteerin makuinen bakteerijuoma
          description: Human readable and user facing name of the product.
        category: 
          '$ref': '#/components/schemas/Category'
        sellPrice: 
          type: integer
          example: 310
          description: Price (in cents) at which the product is on sale.  
        stock: 
          type: integer
          example: 10
          description: Number of items in stock.
    AdminProduct:
      allOf:
        - '$ref': '#/components/schemas/Product'
        - type: object
          required:
            - buyPrice
          properties:
            buyPrice:
              type: integer
              example: 300
              description: Price (in cents) at which the product was bought in.
    Box:
      type: object
      required: 
        - boxBarcode
        - itemsPerBox
        - product
      properties: 
        boxBarcode:
          example: "01880335"
          type: string
          description: |
            Barcode printed on the box, which contains multiple items of a product.
            Used to uniquely identify the box.
        itemsPerBox: 
          type: integer
          example: 18
          description: Number of items contained in one box.
        product:
          '$ref': '#/components/schemas/Product'
...
    Category:
      type: object
      required: 
        - categoryId
        - description
      properties: 
        categoryId: 
          type: integer
          example: 10
          description: Opaque unique identifier of the category.
        description: 
          type: string
          example: Drinks, soda with bottle return
          description: Human readable and user facing description of the product category.
...
  responses:
...
    NotAuthorizedErrorResponse:
      description: Your user lacks the required privilege level to perform this action.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - not_authorized
                  - invalid_token
              message:
                type: string
                example: Not authorized
    NotFoundErrorResponse:
      description: The resource your are requesting does not exist.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - not_found
              message:
                type: string
                example: No product with barcode \'1312767817\' found
    BadRequestErrorResponse:
      description: The request you sent contains invalid values or is missing fields.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - bad_request
                  - invalid_reference
              message:
                type: string
                example: Missing or invalid fields in request
              errors:
                type: array
                items:
                  type: string
                  example: 'boxBarcode should be a numeric 1-14 digit barcode' 
    IdentifierTakenErrorResponse:
      description: Creating a new resource failed, because a resource with the given identifier already exists.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - identifier_taken
              message:
                type: string
                example: Barcode already in use.
paths: 
...
  /api/v1/admin/boxes: 
    get: 
      summary: Get a list of product boxes, optionally filter by item barcode
      operationId: getBoxes
      tags:
        - admin
      security:
        - adminAccessToken: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                itembarcode:
                  description: Only return boxes related to a given barcode
                  type: string
                  example: '4740098010166'
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - boxes
                properties: 
                  boxes: 
                    type: array
                    items: 
                      '$ref': '#/components/schemas/Box'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
    post: 
      summary: Create a new box of products
      operationId: createBox
      tags:
        - admin
      security:
        - adminAccessToken: []
      responses: 
        201: 
          description: A new box created successfully
          content:
            application/json: 
              schema: 
                type: object
                required: 
                  - box
                properties: 
                  box: 
                    '$ref': '#/components/schemas/Box'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        409:
          '$ref': '#/components/responses/IdentifierTakenErrorResponse'
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              required: 
                - boxBarcode
                - itemsPerBox
                - productBarcode
              properties: 
                boxBarcode: 
                  type: string
                  example: '01766752'
                  minLength: 1
                  maxLength: 14
                itemsPerBox: 
                  type: integer
                  example: 18
                  minimum: 1
                productBarcode: 
                  type: string
                  example: '7314660003114'
                  minLength: 1
                  maxLength: 14
  /api/v1/admin/boxes/{boxBarcode}: 
    get: 
      summary: Get the details of a specific product box
      operationId: getBox
      tags:
        - admin
      security:
        - adminAccessToken: []
      parameters: 
        - name: boxBarcode
          description: Uniquely identifying barcode of the box.
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - box
                properties: 
                  box: 
                    '$ref': '#/components/schemas/Box'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
    patch: 
      summary: Update the details of a product box
      operationId: updateBox
      tags:
        - admin
      security:
        - adminAccessToken: []
      parameters: 
        - name: boxBarcode
          description: Uniquely identifying barcode of the box.
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Box updated successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - box
                properties: 
                  box: 
                    '$ref': '#/components/schemas/Box'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              minProperties: 1
              properties: 
                itemsPerBox: 
                  type: integer
                  example: 18
                  minimum: 1
                productBarcode: 
                  type: string
                  example: '7314660003114'
                  minLength: 1
                  maxLength: 14
    delete: 
      summary: Delete a product box
      operationId: deleteBox
      tags:
        - admin
      security:
        - adminAccessToken: []
      parameters: 
        - name: boxBarcode
          description: Uniquely identifying barcode of the box.
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Box deleted successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - deletedBox
                properties: 
                  deletedBox: 
                    '$ref': '#/components/schemas/Box'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
...
  /api/v1/admin/products/{barcode}: 
    get: 
      summary: Get details of a specific product
      operationId: getProduct
      tags:
        - admin
      security:
        - adminAccessToken: []
      parameters: 
        - name: barcode
          description: Uniquely identifying barcode of the product.
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - product
                properties: 
                  product: 
                    '$ref': '#/components/schemas/AdminProduct'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
    patch: 
      summary: Update the details of a specific product
      operationId: updateProduct
      tags:
        - admin
      security:
        - adminAccessToken: []
      parameters: 
        - name: barcode
          description: Uniquely identifying barcode of the product.
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Product updated successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - product
                properties: 
                  product: 
                    '$ref': '#/components/schemas/AdminProduct'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              minProperties: 1
              additionalProperties: false
              properties: 
                name: 
                  type: string
                  example: Muumilimsa
                  minLength: 1
                categoryId: 
                  type: integer
                  example: 10
                  minimum: 0
                buyPrice: 
                  type: integer
                  example: 210
                sellPrice: 
                  type: integer
                  example: 220
                stock: 
                  type: integer
                  example: 8
    delete: 
      summary: Delete a product
      operationId: deleteProduct
      tags:
        - admin
      security:
        - adminAccessToken: []
      parameters: 
        - name: barcode
          description: Uniquely identifying barcode of the product.
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Product deleted successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - deletedProduct
                properties: 
                  deletedProduct: 
                    '$ref': '#/components/schemas/AdminProduct'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
...